+++ lnbits/core/__init__.py
@@ -11,6 +11,7 @@ from .views.fiat_api import fiat_router
 
 # this compat is needed for usermanager extension
 from .views.generic import generic_router
+from .views.lnurl_api import lnurl_router
 from .views.node_api import node_router, public_node_router, super_node_router
 from .views.payment_api import payment_router
 from .views.tinyurl_api import tinyurl_router
@@ -42,6 +43,7 @@ def init_core_routers(app: FastAPI):
     app.include_router(users_router)
     app.include_router(audit_router)
     app.include_router(fiat_router)
+    app.include_router(lnurl_router)
 
 
 __all__ = ["core_app", "core_app_extra", "db"]
